refactor : Canvas에 owner 정보 추가 (#7)
