+++ register/canvas.go
@@ -1,13 +1,12 @@
 package register
 
-import "guestbook/entity"
+import (
+	"guestbook/entity"
+	"guestbook/entity/user"
+)
 
 type Canvas struct {
-	id   entity.ID
-	data map[uint32]LWWRegister
-}
-
-func (llwMap Canvas) IsExist(key uint32) bool {
-	_, exist := llwMap.data[key]
-	return exist
+	id    entity.ID
+	owner user.User
+	data  map[uint32]LWWRegister
 }
